tests/wpt/web-platform-tests/css/css-typed-om/declared-styleMap-accepts-inherit.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Declared styleMap objects accept 'inherit' as a value</title>
<meta name="author" title="Shane Stephens">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
  <div id='element'></div>
  <div id="log"></div>
  <script>
    test(function() {
      element.styleMap.set('width', new CSSKeywordValue('inherit'));
      assert_equals(element.styleMap.get('width').value, 'inherit', 'inherit should be a valid value for styleMap properties');
    });
  </script>
